Allow setting preferences as ux restricted
This deprecates DisabledPreferenceCallback, replacing
it with UxRestrictablePreference. Instead of making
all the views disabled and removing/adding the ripple,
we just apply state_ux_restricted to all the views.
Then the xml customizations can take advantage of that
drawable state, and display themselves as if they were
disabled. This also allows for ux restricted preferences
to look different than disabled preferences.

Still TODO: updating the DeviceDefault customization
of buttons and switches to respect state_ux_restricted,
so that they look correct in the TwoActionPreferences.

Bug: 177233240
Test: Manually, paintbooth settings screen now responds to driving state
Change-Id: I6933fcebf0b538491983aaf02ec52dbef6025f7f