Use getDimensionPixelSize() for sizes
Rather than casting the result of getDimension() to an int, use
getDimensionPixelSize() when fetching a size from a dimen resource.
The former converts the dimension to pixels and rounds down to the
nearest integer, whereas the latter converts the dimenion to pixels
and rounds up or down to the nearest integer, but never rounds down
to zero.

Test: Manual
Bug: 178515570
Change-Id: I5f273e0ddc47aebf407a111ce3adfff3a8913bb9