Make PreferenceFragment layout customizable
Currently in order to customize CarUi's PreferenceFragment, apps need to
pass their custom layout to PreferenceFragment.CarUi theme and then
override onCreateRecyclerView which instantiates CarUiRecyclerView.
The problem is that onCreateRecyclerView expects to return an instance
of androidx Recyclerview which is not available to get from
CarUiRecyclerView anymore.
Instead a method named onCreateCarUiRecyclerView is added that lets apps
instantiate CarUiRecyclerView and not need to access the underlying
androidx Recyclerview.

Bug: 194095014
Test: atest CarUILibUnitTests
Change-Id: Ibfc39c46ecb59bb6f69765e419c2d4d41cd8cc9c