Fix CarUiListItemTest#test_secureListItem_doesntCallOnClickListeners
This test will test if a window could be obscured by another window,
and check if the FLAG_WINDOW_IS_OBSCURED flag would be set when click
on the view.

The window with same uid won't be obscured by another. This CL will
change the test by adding the overlay window from a foreground service
with different UID. So we could make sure if the touch is really
obscured by the overlay window of another package.

Bug: 195938762
Test: atest CarUiListItemTest
Change-Id: Ia2115e047d13233bc17937f867b684131187dcc3